+++ .env.example
@@ -1,0 +1,1 @@
+OPENAI_API_KEY={Your OpenAI API key}
