--- test1.md
@@ -1,1 +1,0 @@
-The key is orange
