--- test2.md
@@ -1,1 +1,0 @@
-The prize is chow
